[APICHANGE] Require objects to stay on one side of the taming membrane. https://codereview.appspot.com/7402056
Previously, an object on the feral side of the taming membrane which
happened to have a prototype chain from a Caja frame would be passed
through without any copying or proxying. We have decided that this is
a hazard since innocent host code might construct a data structure
containing feral references and assume the membrane will protect them;
the membrane now will treat such objects just like host-frame objects.

Additionally, to detect problems with inconsistent use of the membrane
sooner, tame() and untame() will now throw on attempts to tame an object
with a feral twin, or to untame an object with a tame twin. These events
are not problems in themselves but rather indicate inconsistencies: that
the membrane was previously defeated.

This is not a fix for any known security vulnerability; it is intended
to protect containers from accidentally having vulnerabilities in their
use of the taming membrane.

Supporting changes:
* In ES5/3 mode, the host-provided URI policy function now receives a
  URI object which is not implemented by cajoled code.
* Removed unused broken test fixture code from es53-test-taming-tamed.
* Adjusted the '$' function in browser-test-case.js to fit the stricter
  membrane.

Impact:
Host code which took advantage of the previous laxity to pass tame-side
references into the membrane will now fail. Instead, the corresponding
feral reference (untame(obj)) should be used; preferably, code should be
factored so that explicit tame and untame operations are few.

R=ihab.awad@gmail.com