Make Domado PropertyTaming paranoid about non-primitive values. https://codereview.appspot.com/7475044
PropertyTaming provides a number of accessor-generators which are used
only for primitive values. Apply the taming membrane, which should do
nothing as long as the values are actually primitives, but will defend
against the browser unexpectedly returning an object value (exposes
host frame's undefended prototypes) or having some behavior which can be
exploited by the guest setting a gimmick object value.

The Element.prototype.style accessor depended on the old behavior
(getting the feral style object) and has been replaced with a hardcoded
accessor.

Also added a missing semicolon.

R=ihab.awad@gmail.com