Really fix whitelisting of cajaVM.es5ProblemReports. https://codereview.appspot.com/8064045
r5331 was slightly wrong (yes, this is a fix of a fix) in that it added
es5ProblemReports, let it be deleted by the whitelist, and then added it
again. This is largely harmless, but confusing and not warning-clean (it
causes the extra early copy to be reported as deleted).

R=ihab.awad@gmail.com