Update jQuery to 1.9 and remove test cajoling. https://codereview.appspot.com/10087044
Remove the precajoling for jQuery tests because they contain content
which the cajoler rejects, and since ES5/3 mode and the cajoler are
deprecated it is not worth the effort to either loosen the restrictions
or continue to patch the test suite to accept it.

Also, there are now no build dependencies on the jQuery test suite. The
eventual goal is to be able to test with different versions of jQuery by
merely supplying a different test catalog. Currently we still find we
need to patch the test suite itself for strict-mode restrictions.

(jQuery 1.9 is not the latest version any more but it is what I have as
part of this overall change.)

R=felix8a@gmail.com