separate test results https://codereview.appspot.com/10373044
right now, we generate a test report in the same directory
we save test logs, which is inconvenient because the test summary
is an index.html file, making it difficult to use a webbrowser to
look at other files in that directory.

this change moves results, reports, and capture into separate
subdirectories.

R=kpreid2