fix flakiness in cajajs-minor-version-skew tests https://codereview.appspot.com/10576043
The version skew tests rely on making a set of caja js files that
have the wrong version number, so instead of version 4650 the
files would claim to be version 0000 or 4650mmm.

These files are created by string-replacing the current version
4650 with the fake version.

The flakiness happens when there are false positive matches, like
es53-taming-frame.js could have
    'cajoledDate':1372196894650
which is harmless when 4650 becomes 0000, but causes a js syntax
error when 4650 becomes 4650mmm.

This fixes the flakiness by making the version substitution precise.

R=kpreid2