typo check css whitelists https://codereview.appspot.com/10830045
this makes the css pattern generator throw an error if the css
whitelist has a <foo> that isn't defined.

implementing this uncovered a symbol <identifier> that isn't
defined, which is used by counter-increment and counter-reset.

the identifier is a global name, and we were inappropriately
whitelisting counter-reset and counter-increment.  this was
harmless because <identifier> was undefined and ignored,
but it's not harmless now that <identifier> is defined.

so I'm deleting counter-increment and counter-reset from the
whitelist, which is harmless because they didn't work anyway.

R=mikesamuel