fix flakiness in cajajs-bare tests https://codereview.appspot.com/10889045
symptom is the test fails with
  The title shows 0/1 fail 0/1 pass

theory is the test driver is sampling the page's title before the
test has actually run.  cajajs-bare uses an element with the class
"waiting" to signal that it's not done yet, but that was the old
unreliable signalling method, and we're not looking for "waiting"
any more. now, we're looking for elements with the class
"testcontainer", so this fixes that.

R=kpreid2