fix client-side parsing of minified css https://codereview.appspot.com/11470043
The client-side css parser has a couple problems handling minified css.

1. ruleset() always skips the token following the selector, which is
expected to be '{', even if it isn't a '{'. This screws up parsing of
problems like
  @foo{x}y{...}
because the x is a selector and the '}' gets skipped, so the y gets
absorbed into @foo.

2. ruleset() always skips the token following the closing '}'. This
is harmless if the token is whitespace, but screws up but if the
next statement is not separated by whitespace.

This fixes
https://code.google.com/p/google-caja/issues/detail?id=1770

R=mikesamuel