fix parsing of css hacks like '*zoom' https://codereview.appspot.com/12579043
This minified css gets sanitized weirdly:
  p{color:red;*zoom:1;color:blue}
the 'color:blue' gets dropped.

When the css parser expects a declaration but sees '*', it skips one token
and tries to parse another declaration. So in this case, it skips
'*zoom', ':1', ';color', ':blue', and then resynchronizes at the '}'

This CL makes it resynchronize at the next ; or }, which is a bit closer
to what browsers actually do.

R=mikesamuel