Sanitize canvas style property values rather than requiring exact match. https://codereview.appspot.com/12699046
Possibly due to recent changes in the CSS sanitizer, 2D canvas would
not accept rgb() and rgba() colors:
<https://code.google.com/p/google-caja/issues/detail?id=1837>

Now, instead of sanitizing the property and requiring the result to be
identical to the input, we just sanitize the property. If there was a
rationale for doing it the former way, I've forgotten it.

R=mikesamuel@gmail.com