remove dom2property from css whitelist https://codereview.appspot.com/13060045
The css whitelist has an element 'dom2property' that's intended to
map a css name like 'margin-left' to a dom name like 'marginLeft'.

I'm eliminating it, because:

1. It's unnecessary. The name conversion is regular except for
float->cssFloat, and it's unlikely that any other css properties
will be irregular.

2. It's a maintenance hazard. Specifying an incorrect dom2property
will cause mysterious failures. And failing to specify a correct
dom2property may also cause mysterious failures.

There was also legacy support for style->styleFloat instead of cssFloat,
which is for IE<=8, which we're not supporting, so I'm removing the
styleFloat code.

R=kpreid2