upgrade to selenium 2.35.0 https://codereview.appspot.com/13179044
The version of selenium we're using has some flakiness talking to firefox 24 (beta)
Symptom is, occasional timeout when starting a browser.

Upgrading to selenium 2.35.0 seems to fix that.

I'm also upping the timeout for es53 scan to 5 minutes, because it sometimes
times out when I'm doing something else at the same time.

(this CL doesn't include the update to selenium-standalone.jar because
codereview barfs on huge files like that)

R=kpreid2