fix Closure Compiler issues in html-css-sanitizer https://codereview.appspot.com/13258043
This fixes the css sanitizer to be Closure friendly and replaces the
minified version with a closured version. This reduces the size from
76k to 52k uncompressed, 24k to 16k compressed.

The fix is all about how Closure treats a.foo differently from
a['foo']. These particular foos are external linkage names, so
I changed them to a['foo'] so Closure won't rename them.

The change in RhinoTestBed is because the Closured js will emit
code like
   if (condition) /foo/.test(bar)
and our java JS parser barfs on the regexp in that position.
The barf is a known problem, it's complicated to fix, but it
turns out we don't really need to parse js in RhinoTestBed,
so that's what the change is.

R=ihab.awad