fix rare flakiness in domado-events testRepeatedHandlers https://codereview.appspot.com/13414043
This addresses https://code.google.com/p/google-caja/issues/detail?id=1779

Symptom is we expect
 mousedown,1,mousedown,3,mouseup,1,mouseup,3
but we get
 mouseup,1,mouseup,3

I haven't been able to reproduce it, but I think what's happening is
that the webdriver click() puts a small delay between mousedown and
mouseup events, and in that interval we happen to run the testcase,
which installs mousedown and mouseup handlers, and then sees
results from just the mouseup events firing.

This fix just reduces the test to mouseup handlers, since it's
intended to test that repeated calls to addEventListener do
the expected thing, and there's nothing special about testing
both mouseup and mousedown.

R=kpreid2