remove css-history-sniff defense code https://codereview.appspot.com/33640043
Modern browsers all implement history-sniff protection in basically
the same way Caja does, so this CL deletes the code in Caja.

This CL is for the es53 branch, because the java css sanitizer
is deleted in trunk. I'll port this to trunk in a separate CL.

The js function sanitizeCssSelectors returns two lists:
non-history-sensitive and history-sensitive selectors.
Since there seem to be no external users of sanitizeCssSelectors,
I'm deliberately breaking compatibility by renaming the function
to santizeCssSelectorList and having it return just a single list.

css-stylesheet-tests.js has some error messages added. This is because
I moved an assertNoErrors() statement in CssRewriterTest.
Previously, CssRewriterTest only checked error messages if
css-stylesheet-tests declared a "messages" element.
I felt uncomfortable about swallowing unexpected errors,
so I modified it to always check errors, whether or not "messages"
is specified.

test-domado-special-guest had a testcase that checks that history
sniffing doesn't work. In principle, we could keep this test and
check the browser's history-sniffing defense. In practice, this
was a really complicated change, because the imaginary computed
style returned by Caja was nontrivially different from the
imaginary computed style returned by browsers. After spending
too long trying to fix the test, I decided to just delete it
instead, since it's not really checking Caja code.

R=kpreid2