Add leniency in taming for better Google Visualization API support. https://codereview.appspot.com/57170046
Loading google.visualization with options such as
    {packages:['orgchart']}
will result in a stub "google.visualization.PieChart" which is not a
function, causing markCtor to fail. To avoid failures like this, skip
taming any object which is expected to be a ctor but is not a function.

Supporting changes:
In the test server, request disabling of browser XSS protection.
This sometimes interferes with the generic-host-page since it
evaluates scripts specified in the URL, and could plausibly hide
an actual unsandboxed execution bug in Caja from the tests.

R=ihab.awad@gmail.com