Properly render FF 30 Nightly stack traces. https://codereview.appspot.com/69820043
FF 30 Nightly stack traces are more informative for nested
evaluation. But the extra info does not correspond well to the
patterns we look for. Extend the lineColPattern into a prioritized
sequence of patterns, in the same way that framePatterns is a
prioritized sequence. Add a lineColPattern specialized for the FF30
Nightly nested eval pattern.

R=kpreid.switchb.org@gmail.com