Work around Firefox 33 regexp performance bug in CSS lexer. https://codereview.appspot.com/157080044
In Firefox 33, regexp operations on strings which can be represented
as Latin-1 characters only may be very slow or hung. Work around
this for our CSS lexer by inserting a funny character temporarily.

This change is based on a patch provided by James Keane, with
additional explanatory text and rearrangement of the existing code.

The browser bug is expected to be fixed in Firefox 34.

https://code.google.com/p/google-caja/issues/detail?id=1941
https://bugzilla.mozilla.org/show_bug.cgi?id=1081175

R=erights@gmail.com