Have test links link more directly to github test262 pages. https://codereview.appspot.com/199650043
The old heuristic for dereferencing test: identifiers has not worked
for a long time. Change these from test ids to more direct links into
the github test262 repository. For brevity, still recognize when a
link starts with "test/", in which case prepend the github link
prefix.

Two tests use the full url because the tests have been removed from
github, and so we link to an historical version.

R=kpreid.switchb.org@gmail.com