Repair Safari Throw-Thaw Vulnerability https://codereview.appspot.com/214110043
Webkit bug https://bugs.webkit.org/show_bug.cgi?id=141878 is that
throwing a frozen object unfreezes it in a way that creates an
opportunity for a capability leak. This CL extends the existing test
and adds a repair to make things safe from this bug on Safari.

Note: The first snapshot does none of this, but just merges the
outstanding changes from https://codereview.appspot.com/202030043/ and
https://codereview.appspot.com/202040043/ . The actual content of this
CL will be the differential from those.

R=jasvir@gmail.com